Add tests for writing sequences to a handle
To support the uncommon but important use case of writing individual
sequence records to an open file handle, this commit adds unit tests for
a proposed change to the `write_sequences` interface that supports that
use case. The new interface includes two optional arguments and an
optional second output that allows the caller to work with the file
handle opened to write sequences or pass its own file handle.