Refactor parse logic and use read/write interface
Refactors logic from an original for loop into its own `parse_sequence`
function, adds tests for this new function, and updates the body of the
`run` function to use this function inside the for loop. This commit
also replaces the Bio.SeqIO read and write functions with the new
`read_sequences` and `write_sequences` functions. These functions
support compressed input and output files based on the filename
extensions.